fix: resolve critical pipeline and notebook issues
Fix three critical bugs identified in code review:

1. Notebook cell corruption in refactor script
   - Replace split('\n') with splitlines(keepends=True) in 6 locations
   - Preserves newline terminators in Jupyter cells
   - Fixes syntax errors and import concatenation issues

2. Logging handler duplication in pipeline runner
   - Add deduplication logic before adding handlers
   - Prevents double-logging in repeated pipeline runs
   - Clears existing generic handlers to avoid buildup

3. Type mismatch in stage registry
   - Change stage_number from int to float
   - Supports sub-stage numbering (7.5, 7.6, 7.7)
   - Maintains semantic stage ordering

Also includes:
- Regenerated ML_Pipeline.ipynb with corrected script
- Added entropy features and deflated Sharpe ratio validation
- Updated CV runner and feature engineering

Verified:
- 30/30 integration tests passed
- Notebook format validation passed
- Stage registry accepts float stage numbers

+++ src/pipeline/stage_registry.py
@@ -18,7 +18,7 @@ class PipelineStage:
     description: str = ""
     required: bool = True
     can_run_parallel: bool = False
-    stage_number: int = 0
+    stage_number: float = 0.0
 
     def __post_init__(self):
         """Validate stage definition."""
